fix(task_queue): cancellation safety and priority range validation
Issue 1 — cancellation safety in TaskQueue::acquire:
Add AcquireGuard RAII guard that tracks acquisition phase
(ProjectWait / GlobalWait). If the future is dropped at either
.await point (e.g. client disconnect), Drop decrements all
incremented counters and releases any already-acquired project
permit, preventing permanent capacity leaks and false
max_queue_size errors.

Issue 2 — priority range validation:
Add MAX_TASK_PRIORITY = 2 constant to task_runner. Validate
priority <= 2 in enqueue_task, enqueue_task_background
(task_routes.rs) and DefaultExecutionService::validate_request
(execution.rs). Requests with priority > 2 are rejected with
a 400 BadRequest error, preventing scheduler-level starvation.

Also extract task_queue test module to task_queue_tests.rs to
keep task_queue.rs under the 800-line hard limit.
